requirements.txt
streamlit
pandas
transformers
torch
sentencepiece
plotly
Pillow
google-generativeai